odo-maven-plugin/src/test/java/com/lordofthejars/odo/maven/OdoComponentDeleteMojoTest.java
package com.lordofthejars.odo.maven;

import com.lordofthejars.odo.core.Odo;
import com.lordofthejars.odo.testbed.junit5.OdoExecutorStubInjector;
import com.lordofthejars.odo.testbed.odo.OdoExecutorStub;
import org.apache.maven.plugin.MojoExecutionException;
import org.apache.maven.plugin.MojoFailureException;
import org.apache.maven.project.MavenProject;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import java.io.File;
import java.util.HashMap;
import java.util.Map;

import static com.lordofthejars.odo.testbed.assertj.OdoExecutorAssertion.assertThat;
import static org.mockito.Mockito.when;

@ExtendWith({MockitoExtension.class, OdoExecutorStubInjector.class})
public class OdoComponentDeleteMojoTest {
    @Mock
    MavenProject project;

    @Test
    public void testMojoBehavior(OdoExecutorStub odoExecutorStub) throws MojoExecutionException, MojoFailureException {
        // Given
        OdoComponentDeleteMojo odoComponentDeleteMojo = new OdoComponentDeleteMojo();
        Odo odo = new Odo(odoExecutorStub);

        when(project.getBasedir()).thenReturn(new File("/tmp/foodir"));
        Map<String, String> componentDeleteConfiguration = new HashMap<>();
        componentDeleteConfiguration.put("project", "fooproject");

        odoComponentDeleteMojo.deleteComponent = componentDeleteConfiguration;
        odoComponentDeleteMojo.project = project;
        odoComponentDeleteMojo.componentName = "mycomponent";

        odoComponentDeleteMojo.odo = odo;

        // When:
        odoComponentDeleteMojo.execute();

        // Then:
        assertThat(odoExecutorStub).hasExecuted("odo component delete mycomponent --project fooproject");
    }
}